up to quick view

+++ app/Http/Controllers/pageSettingController.php
@@ -48,7 +48,7 @@ class pageSettingController extends Controller
     }
 
     public function showSlider(){
-        $slider= homeSlider::all();
+        $slider= homeSlider::orderBy('position', 'ASC')->get();
         return view('admin.setting.slider',compact('slider'));
     }
 
@@ -76,6 +76,11 @@ class pageSettingController extends Controller
        $slider->sub_y = $request->sub_y;
        $slider->desc_color = $request->desc_color;
        $slider->desc_y = $request->desc_y;
+       $position_count = count(homeSlider::all());
+        if(empty($position_count)){
+            $position_count = 0;
+        }
+        $slider->position = $position_count;
        $slider->save();
        return response()->json($slider);
     }
@@ -116,15 +121,69 @@ class pageSettingController extends Controller
     }
     public function deleteSlider($id){
         $slider = homeSlider::find($id);
-        $remove_file = public_path().'/slider/'. $slider->slider_image;
-        if (file_exists($remove_file)) {
-            unlink($remove_file);
+        if($slider->slider_image !=null){
+            $remove_file = public_path().'/slider/'. $slider->slider_image;
+            if (file_exists($remove_file)) {
+                unlink($remove_file);
+            }
         }
-        $slider->delete();
+
+        $sliderAll = homeSlider::all();
+        if($slider->position != count($sliderAll)-1){
+            $result = homeSlider::where('position','>',$slider->position)->orderBy('position', 'ASC')->get();
+            $x= $slider->position;
+            foreach($result as $row){
+                $slider_Update = homeSlider::find($row->id);
+                $slider_Update->position = $x;
+                $slider_Update->save();
+                $x++;
+            }
+        }
+        $slider1 = homeSlider::find($id);
+        $slider1->delete();
         return response()->json("200");
     }
+
+    public function dropSlider($index,$id){
+        $slider = homeSlider::find($id);
+        $storeslider = array();
+        if($slider->position > $index){
+          for ($i= $index; $i < $slider->position; $i++) { 
+              $num = $i;
+            $slider_drop = homeSlider::where('position',$i)->first();
+            $slider_drop->position = $num+1;
+            //$slider_drop->save();
+            $storeslider[] = $slider_drop;
+          }
+          foreach($storeslider as $row){
+            $slider_Update = homeSlider::find($row->id);
+            $slider_Update->position = $row->position;
+            $slider_Update->save();
+          }
+
+        }else{
+            for ($i= $index; $i > $slider->position; $i--) { 
+                $num = $i;
+              $slider_drop1 = homeSlider::where('position',$i)->first();
+              $slider_drop1->position = $num-1;
+             // $slider_drop->save();
+             $storeslider[] = $slider_drop1;
+            } 
+            foreach($storeslider as $row){
+                $slider_Update = homeSlider::find($row->id);
+                $slider_Update->position = $row->position;
+                $slider_Update->save();
+              }
+        }
+        $slider1 = homeSlider::find($id);
+        $slider1->position = $index;
+        $slider1->save();
+        return response()->json($storeslider); 
+       
+    }
+
     public function homeLayout(){
-        $layout = home_product_layout::all();
+        $layout = home_product_layout::orderBy('position', 'ASC')->get();
         $product = product::all();
         $category = category::all();
         return view('admin.setting.layout',compact('layout','product','category'));
@@ -138,6 +197,11 @@ class pageSettingController extends Controller
         }else{
             $layout->category_id = collect($request->category)->implode(',');
         }
+        $position_count = count(home_product_layout::all());
+        if(empty($position_count)){
+            $position_count = 0;
+        }
+        $layout->position = $position_count;
         $layout->save();
         return response()->json($request); 
     }
@@ -191,8 +255,57 @@ class pageSettingController extends Controller
     
     public function DeleteLayout($id){
         $layout = home_product_layout::find($id);
-        $layout->delete();
+        $layoutAll = home_product_layout::all();
+        if($layout->position != count($layoutAll)-1){
+            $result = home_product_layout::where('position','>',$layout->position)->orderBy('position', 'ASC')->get();
+            $x= $layout->position;
+            foreach($result as $row){
+                $layout_Update = home_product_layout::find($row->id);
+                $layout_Update->position = $x;
+                $layout_Update->save();
+                $x++;
+            }
+        }
+        $layout1 = home_product_layout::find($id);
+        $layout1->delete();
         return response()->json("200");
     }
     
+    public function dropLayout($index,$id){
+        $layout = home_product_layout::find($id);
+        $storeLayout = array();
+        if($layout->position > $index){
+          for ($i= $index; $i < $layout->position; $i++) { 
+              $num = $i;
+            $layout_drop = home_product_layout::where('position',$i)->first();
+            $layout_drop->position = $num+1;
+            //$layout_drop->save();
+            $storeLayout[] = $layout_drop;
+          }
+          foreach($storeLayout as $row){
+            $layout_Update = home_product_layout::find($row->id);
+            $layout_Update->position = $row->position;
+            $layout_Update->save();
+          }
+
+        }else{
+            for ($i= $index; $i > $layout->position; $i--) { 
+                $num = $i;
+              $layout_drop1 = home_product_layout::where('position',$i)->first();
+              $layout_drop1->position = $num-1;
+             // $layout_drop->save();
+             $storeLayout[] = $layout_drop1;
+            } 
+            foreach($storeLayout as $row){
+                $layout_Update = home_product_layout::find($row->id);
+                $layout_Update->position = $row->position;
+                $layout_Update->save();
+              }
+        }
+        $layout1 = home_product_layout::find($id);
+        $layout1->position = $index;
+        $layout1->save();
+        return response()->json($storeLayout); 
+       
+    }
 }
